backend: move discovery account numbering
